requirements.txt
python-telegram-bot==21.6
psutil==6.1.0
